[Refactor] Overwrite original example_cross_entro.py with optimized version
Replace the separate cross_entropy.py with direct overwrite of the
original file per submission preference.
